+++ config.el
@@ -90,19 +90,27 @@ END
 ;; значит, если продолжать в этом направлении, надо заменить пару
 ;; (regexp rep) на список пар
 ;; лучше использовать el-patch
-(defun eab/patch-this-code (func-name regexp rep &optional lex)
-  (let ((overriding-terminal-local-map (make-sparse-keymap)))
-    (eval
-     (read
-      (replace-regexp-in-string
-       regexp
-       rep
-       (save-window-excursion
-	 (find-function-do-it func-name nil 'switch-to-buffer)
-	 (let ((bgn (point)))
-	   (forward-sexp)
-	   (let ((end (point)))
-	     (buffer-substring-no-properties bgn end)))))) lex)))
+(require 'cl-macs)
+(cl-defun eab/patch-this-code (func-name rpairs &optional &key lexical native)
+  (let* ((overriding-terminal-local-map (make-sparse-keymap))
+	 (func-string (save-window-excursion
+			(find-function-do-it func-name nil 'switch-to-buffer)
+			(let ((bgn (point)))
+			  (forward-sexp)
+			  (let ((end (point)))
+			    (buffer-substring-no-properties bgn end)))))
+	 (func-code (read
+		     (seq-reduce
+		      (lambda (string regexp-replacement-pair)
+			(replace-regexp-in-string
+			 (car regexp-replacement-pair)
+			 (cdr regexp-replacement-pair)
+			 string))
+		      rpairs
+		      func-string))))
+    (if native
+	(native-compile (eval func-code lexical))
+      (eval func-code lexical))))
 
 (defun revert-all-buffers ()
   "Refreshes all open buffers from their respective files."