+++ lisp/eab-lisp.el
@@ -1,0 +1,77 @@
+;;; eab-lisp.el --- let over lambda -*- lexical-binding: t -*-
+
+(cl-defun mkstr (&rest args)
+  (with-output-to-string
+    (cl-dolist (a args) (princ a))))
+
+(cl-defun symb (&rest args)
+  (values (intern (apply #'mkstr args))))
+
+(defun group (source n)
+  (if (zerop n) (error "zero length"))
+  (labels ((rec (source acc)
+             (let ((rest (nthcdr n source)))
+               (if (consp rest)
+                   (rec rest (cons
+                               (subseq source 0 n)
+                               acc))
+                   (nreverse
+                     (cons source acc))))))
+    (if source (rec source nil) nil)))
+
+(defun flatten (x)
+  (labels ((rec (x acc)
+             (cond ((null x) acc)
+                   ((atom x) (cons x acc))
+                   (t (rec
+                        (car x)
+                        (rec (cdr x) acc))))))
+    (rec x nil)))
+
+(defun g!-symbol-p (s)
+  (and (symbolp s)
+       (> (length (symbol-name s)) 2)
+       (string= (substring (symbol-name s) 0 2)
+                "g!")))
+
+(cl-defmacro defmacro/g! (name args &body body)
+  (let ((syms (remove-duplicates
+                (remove-if-not #'g!-symbol-p
+                               (flatten body)))))
+    `(defmacro ,name ,args
+       (let ,(mapcar
+               (lambda (s)
+                 `(,s (gensym ,(subseq
+                                 (symbol-name s)
+                                 2))))
+               syms)
+         ,@body))))
+
+(defun o!-symbol-p (s)
+  (and (symbolp s)
+       (> (length (symbol-name s)) 2)
+       (string= (substring (symbol-name s) 0 2)
+                "o!")))
+
+(defun o!-symbol-to-g!-symbol (s)
+  (symb "g!"
+        (subseq (symbol-name s) 2)))
+
+(defmacro! dlambda (&rest ds)
+  `(lambda (&rest ,g!args)
+     (case (car ,g!args)
+       ,@(mapcar
+           (lambda (d)
+             `(,(if (eq t (car d))
+                  t
+                  (list (car d)))
+               (apply (lambda ,@(cdr d))
+                      ,(if (eq t (car d))
+                         g!args
+                         `(cdr ,g!args)))))
+           ds))))
+
+(funcall (let ((count 0))
+	   (dlambda
+	    (:inc () (incf count))
+	    (:dec () (decf count)))) :inc)